feat: 모집 기능에 사용자 권한 및 소유권 검증 로직 추가 (#127)

+++ com.trillionares.tryit.product/src/main/java/com/trillionares/tryit/product/application/service/RecruitmentService.java
@@ -1,5 +1,6 @@
 package com.trillionares.tryit.product.application.service;
 
+import com.trillionares.tryit.product.domain.client.AuthClient;
 import com.trillionares.tryit.product.domain.common.json.JsonUtils;
 import com.trillionares.tryit.product.domain.model.recruitment.Recruitment;
 import com.trillionares.tryit.product.domain.model.recruitment.type.RecruitmentStatus;
@@ -31,11 +32,16 @@ public class RecruitmentService {
 
     private final RecruitmentRepository recruitmentRepository;
     private final KafkaTemplate<String, String> kafkaTemplate;
+    private final AuthClient authClient;
 
 
     @Transactional
-    public RecruitmentIdResponse createRecruitment(CreateRecruitmentRequest request) {
+    public RecruitmentIdResponse createRecruitment(CreateRecruitmentRequest request, String username, String role) {
+        UUID userId = authClient.getUserByUsername(username).getData().getUserId();
+        validatePermission(role);
+
         Recruitment recruitment = Recruitment.builder()
+                .userId(userId)
                 .productId(request.productId())
                 .recruitmentTitle(request.title())
                 .recruitmentDescription(request.description())
@@ -53,9 +59,14 @@ public class RecruitmentService {
 
     @Transactional
     public RecruitmentIdResponse updateRecruitment(UUID recruitmentId,
-                                                   UpdateRecruitmentRequest request) {
+                                                   UpdateRecruitmentRequest request, String username, String role) {
+        UUID userId = authClient.getUserByUsername(username).getData().getUserId();
+        validatePermission(role);
+
         Recruitment recruitment = recruitmentRepository.findById(recruitmentId)
-                .orElseThrow(() -> new RuntimeException(""));
+                .orElseThrow(() -> new RuntimeException("Not Found Recruitment"));
+
+        validateOwnership(recruitmentId, userId);
 
         recruitment.updateRecruitment(request.title(), request.description(), request.startTime(),
                 request.during(), request.endTime(), request.maxParticipants());
@@ -66,10 +77,14 @@ public class RecruitmentService {
     }
 
     @Transactional
-    public RecruitmentIdResponse deleteRecruitment(UUID recruitmentId) {
-        Recruitment recruitment = recruitmentRepository.findById(recruitmentId)
-                .orElseThrow(() -> new RuntimeException(""));
+    public RecruitmentIdResponse deleteRecruitment(UUID recruitmentId, String username, String role) {
+        UUID userId = authClient.getUserByUsername(username).getData().getUserId();
+        validatePermission(role);
 
+        Recruitment recruitment = recruitmentRepository.findById(recruitmentId)
+                .orElseThrow(() -> new RuntimeException("Not Found Recruitment"));
+
+        validateOwnership(recruitmentId, userId);
         // BaseEntity 구현 후 soft delete 로 변경
         recruitmentRepository.delete(recruitment);
 
@@ -87,11 +102,18 @@ public class RecruitmentService {
         return recruitmentRepository.getRecruitmentList(pageable);
     }
 
+    @Transactional
     public UpdateRecruitmentStatusResponse updateRecruitmentStatus(UUID recruitmentId,
-                                                                   UpdateRecruitmentStatusRequest request) {
-        Recruitment recruitment = recruitmentRepository.findById(recruitmentId)
-                .orElseThrow(() -> new RuntimeException(""));
+                                                                   UpdateRecruitmentStatusRequest request,
+                                                                   String username,
+                                                                   String role) {
+        UUID userId = authClient.getUserByUsername(username).getData().getUserId();
+        validatePermission(role);
 
+        Recruitment recruitment = recruitmentRepository.findById(recruitmentId)
+                .orElseThrow(() -> new RuntimeException("Not Found Recruitment"));
+
+        validateOwnership(recruitmentId, userId);
         recruitment.updateStatus(request.status());
 
         recruitmentRepository.save(recruitment);
@@ -140,7 +162,7 @@ public class RecruitmentService {
     public RecruitmentExistAndStatusDto isExistRecruitmentById(UUID recruitmentId) {
         Optional<Recruitment> recruitment = recruitmentRepository.findByRecruitmentId(recruitmentId);
 
-        if(!recruitment.isPresent() || recruitment.isEmpty() || recruitment == null) {
+        if (!recruitment.isPresent() || recruitment.isEmpty() || recruitment == null) {
             return RecruitmentExistAndStatusDto.of(false, "NOT_FOUND");
         }
 
@@ -160,4 +182,19 @@ public class RecruitmentService {
                 return RecruitmentExistAndStatusDto.of(false, "NOT_FOUND");
         }
     }
+
+    private void validatePermission(String role) {
+        if (!(role.contains("ADMIN") || role.contains("COMPANY"))) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+    }
+
+    private void validateOwnership(UUID recruitmentId, UUID userId) {
+        UUID ownerId = recruitmentRepository.findOwnerIdByRecruitmentId(recruitmentId)
+                .orElseThrow(() -> new RuntimeException("Recruitment not found."));
+        if (!ownerId.equals(userId)) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+    }
+
 }
